pyproject.toml
[tool.poetry]
name = "hierarchicaldata"
version = "0.1.0"
description = ""
authors = ["Triston Reeves <triston.reeves123@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
